Redirection and simple Caching
There is nothing in the PSR-7 ResponseInterface to get the actual uri
that was returned, if a request uri returns a redirect response, that
should actually be the response and client implementations need to take
care of redirection themselves. So the HttpLoader now takes care of
redirects. To be able to get the original requested uri and the final
effective uri, implement the RequestResponseAggregate.

Remove guzzle7 adapter as I found that the guzzle client itself also
implements the PSR-18 ClientInterface.

Also introduce some simple caching functionality, so you don't have to
load pages again and again when in development mode for a crawler.