Fix reading response to string
When you get the response body from a PSR-7 response object and use
getContents() to get the whole body as a string, it looks like it
empties the response body in the object.
As for caching we need to get it as a string, add a new Response
instance cloned from the original, adding the body again, to the
RequestResponseAggregate.