Add a new type of step group
The previous group is renamed to ParallelGroup, because all the steps
in it are invoked virtually in parallel with the same input from the
previous step.
The new SequentialGroup uses the output of the previous step as the
input for the next step, so they can only be run in sequence.