Improve invoke methods and repeating steps
invoke method now always is a Generator, because with a plain returning
method it's not possible to don't return anything at all. Or let's say
to differentiate if it actually knowingly returned null or not. With
Generators it's possible that a Step doesn't produce any output at all.
Further finished functionality for self repeating step and test it.