Add loop functionality
A LoopStep wraps an implementation of the StepInterface and calls it in
a loop with its own output as long as it has output. There has to be a
limit of loop iterations to avoid infinite loops which by default is
1000. Also you can manipulate the Step's output before giving it as
input to the next iteration using the transformOutputToInput method. It
either takes a Closure or you can also use a StepInterface there.
Which is very useful for pagination: you can loop the Http step to
load a page and use a GetLink step in transforming the output to input
to get the url of the next page as next input.
When the transformer has no output the loop is also stopped, which is
also important for the pagination example.

Also change that initial input for the crawler is provided via a
separate input method and not as argument for the run method. Which
makes setting up the crawler steps more readable, because it looks
confusing when the first input visually is at the end.