Improve step groups
Go back to just one type of group which is the same as the ParallelGroup
was. If something like a sequential group would be needed again I'd
just call it sequence or chain or something. Also add a static method to
the Crawler class which allows for this handy usage:

$crawler->addStep(
    Crawler::group()
        ->addStep(Html::querySelector('.foo'))
        ->addStep(Html::querySelector('.bar'))
        ...
);

Remove the GroupInterface and the addGroup method. A Group (just like a
Loop) should just be used like a Step by the Crawler, so just implement
the StepInterface.