Require PHP 8.1 and improve Http Steps for looping
When users extend the Http step they could manipulate class properties
for the request within the invoke method. When looping such a step that
can have undesirable side effects. So I think step class properties set
via the constructor should always be readonly. As this is only available
in PHP 8.1, but really helpful for this issue I upgrade the required
PHP version to 8.1.