Improve not passing on outputs to next step
Rename dontYield to dontCascade and add cascades method to steps to get
the info if a step should cascade it's output. Chose that because a step
should still yield it's output so methods like updateInputUsingOutput
still get it, but the Crawler (and Groups) don't pass on (= cascade) the
output to the next step.

Also start the code style to always have a blank line between lines with
assignments and simple calls.