Change keep looping without output behaviour
Found that I need it also when there was already output in a previous
loop iteration. But I think always having the last call without the
step yielding any output will be weird to users. So add a param to the
withInput method and also an extra method to set this param to only
have this behaviour when the user needs it.