First step to refactor DOM parsing
Got a better idea for steps parsing a DOM, html or xml. Both can be done
with the Symfony DOM Crawler and XPath or CSS selectors, so most code
is shared. You can now select multiple things within one step and even
iterate over a list, or get first or last elements and extract an array
of data from that.
Replacing all the existing HTML parsing steps will be a bit more work,
so this is just the first preparation step in replacing it.