Add/improve int.-tests and getting Http responses
Running the integration tests with the test server had problems when
running tests in parallel, so split integration tests and run them
separately, not in parallel. Also add them to the git pre-commit hook.

Add a new integration test to test the step grouping feature.

Replace getting the body string from a response everywhere with a new
central helper method that rewinds the stream before and after getting
the response. Also replace variable naming from before renaming
RequestResponseAggregate to RespondedRequest.

Last but not least, change the Http loading step to only yield output
when the response is not null.