Get unique outputs from steps
Calling the uniqueOutputs() method on any step will cause that it only
cascades unique outputs. If you pass a key to the method, and the
outputs are arrays or objects having that key/property, it will make
outputs unique based on that key.

Besides that add a lot of blank lines again according to the code style
I practiced recently and add a lot of final keywords for methods in the
Step class.