Implement Group setResultKey and addKeysToResult
Both methods work for Groups only when combineToSingleOutput is used
(and was called before), otherwise they'll throw an Exception.
setResultKey, will add the whole combined output as value with the
defined key. addKeysToResult adds keys from the combined output to the
result.
Also change how the first param of addStep() works in groups: when
its a string, don't set the result key, but use it as key in the
combined output.