Fix Generators usage
Previously the use of Generators actually didn't make a lot of sense,
because the outputs of one step were only iterated and passed on to the
next step, after the current step was invoked with all its inputs. That
makes steps with a lot of inputs bottlenecks and causes bigger memory
consumption. So, change the crawler to immediately pass on outputs of
one step to the next step if there is one.

Also simplify how unique step output works. The crawler class doesn't
need to know if a step should produce unique output and handle it,
instead the step saves the unique keys in a class property until it is
reset by the Crawler (after a full run).
This also makes unique output from loops a lot simpler, which now also
works without cascadeWhenFinished.