Make Dom::extract() work with single selector
The `extract()` method of the `Html` and `Xml` (children of `Dom`) steps
now also works with a single selector instead of an array with a
mapping. Sometimes you'll want to just get a simple string output e.g.
for a next step, instead of an array with mapped extracted data.