Fix selectors not matching anything
When a `DomQuery` (CSS selector or XPath query) doesn't match anything,
its `apply()` method now returns `null` (instead of an empty string).
When the `Html(/Xml)::extract()` method is used with a single, not
matching selector/query, nothing is yielded. When it's used with an
array with a mapping, it yields an array with null values. If the
selector for one of the methods `Html(/Xml)::each()`,
`Html(/Xml)::first()` or `Html(/Xml)::last()` doesn't match anything,
that's not causing an error any longer, it just won't yield anything.