Get absolute links when extracting data
In order to be able to also get absolute links when using the
`extract()` method of Dom steps, the abstract `DomQuery` class now has
a method `toAbsoluteUrl()`. The Dom step will automatically provide the
`DomQuery` instance with the base url, presumed that the input was an
instance of the `RespondedRequest` class and resolve the selected value
against that base url.