Add option to use headless browser with HttpLoader
Add chrome-php/chrome composer dependency and simple option to use a
headless chrome browser with the HttpLoader. So this is enough to get
HTML after executing javascript in the browser. For more sophisticated
tasks better not extend the HttpLoader further, but instead maybe
create a separate Loader and/or Steps.