Add new HTTP::crawl() step
It's for classic crawling all pages of a website. By default it follows
all links on the same host, optionally all on same domain, including
subdomains. You can start with any HTML document or optionally with one
or more sitemaps. You can tell it to only load pages to a certain depth.
You can filter URLs to load, by the path or even check the link element
where it was found to decide if it should follow. You can also decide to
load all pages, but yield only the responses matching your criteria
(path, link element, custom URL check) to the next step.
Besides that, improve the onError callback in the HttpLoader, fix a
 forgotten namespace in a test file and update the url and robots-txt
 package dependencies.