Improve Politeness features
Get rid of the traits, because traits are just stupid and also having
the PoliteHttpLoader makes everything just more complex. Converted the
traits to dependencies of the HttpLoader class => RobotsTxtHandler and
Throttler. The HttpLoader gets defaults of those two classes. The
RobotsTxtHandler checks if you're using an instance of BotUserAgent or
a normal UserAgent and it will respect robots.txt only when using a
BotUserAgent. The Throttler by default throttles to waiting between 0.5
and 1.0 of the previous response time and at least 0.1 seconds, but can
easily be configured differently.