Handling 429 - Too Many Requests
In case of a 429 (Too Many Requests) response, the `HttpLoader` now
automatically waits and retries. By default, it retries twice and waits
10 seconds for the first retry and a minute for the second one. In case
the response also contains a `Retry-After` header with a value in
seconds, it complies to that. Exception: by default it waits at max `60`
seconds (you can set your own limit if you want), if the `Retry-After`
value is higher, it will stop crawling. If all the retries also receive
a `429` it also throws an Exception and therefor stops crawling.

Besides that also fixed handling cookies for localhost or other hosts
without a registrable domain name. And also removed the logger from the
Throttler, as it doesn't log anything.