Controll what to do in case of error responses
New methods `stopOnErrorResponse()` and `yieldErrorResponses()` that can
be used with `Http` steps. By calling `stopOnErrorResponse()` the step
will throw a `LoadingException` when a response has a 4xx or 5xx status
code. By calling the `yieldErrorResponse()` even error responses will be
yielded and passed on to the next steps.

The latter actually was the default behavior until now, but I think most
people would either like to just ignore error responses or the whole
crawler to fail/stop, so I changed this.