New functionality to paginate
There is the new `Paginate` child class of the `Http` step class (easy
access via `Http::get()->paginate()`). It takes an instance of the
`PaginatorInterface` and uses it to iterate through pagination links.
There is one implementation of that interface, the
`SimpleWebsitePaginator`. The `Http::get()->paginate()` method uses it
by default, when called just with a CSS selector to get pagination
links. Paginators receive all loaded pages and implement the logic to
find pagination links. The paginator class is also called before sending
a request, with the request object that is about to be sent as an
argument (`prepareRequest()`). This way, it should even be doable to
implement more complex pagination functionality. For example when
pagination is built using POST request with query strings in the request
body.