Add default timeouts for the default guzzle client
The default values are 10 seconds for connect_timeout and 60 seconds for
timeout. Also add the option to provide config options for the default
guzzle client created inside the `HttpLoader`.

Further, removed the `httpClient()` method in the `HttpCrawler` class.
If you want to provide your own HTTP client, implement a custom `loader`
method passing your client to the `HttpLoader` instead.