Groups: only combined outputs + new outputKey method
Group steps can now only produce combined outputs, as previously done
when `combineToSingleOutput()` method was called. The method is removed.
This was done because I don't really see any actual use case for a group
with all the steps yielding separate outputs.

Using the new `outputKey()` method, a step will convert non array output
to an array and use the key provided as argument to this method as array
key for the output value.