Improve adding data to final Result objects
Add new step methods `addToResult()` and `addLaterToResult()`.
`addToResult()` is a single replacement for `setResultKey()` and
`addKeysToResult()` (which are removed) that can be used for array and
non array output.
`addLaterToResult()` is a new method that does not create a Result
object immediately, but instead adds the output of the current step to
all the Results that will later be created originating from the current
output.

New step method `createsResult()`, so you can differentiate if a step
creates a Result object, or just keeps data to add to results later
(new `addLaterToResult()` method). Primarily relevant for library
internal use.

`getResultKey()` is also removed with `setResultKey()`. It's removed
without replacement, as it doesn't really make sense any longer.