Improve cache implementation
Breaking: The `FileCache` now also respects the `ttl` (time to live)
argument and by default it is one hour (3600 seconds). If you're using
the cache and expect the items to live (basically) forever, please
provide a high enough value for default the time to live. When you try
to get a cache item that is already expired, it (the file) is
immediately deleted.

Also refactored the implementation internally. The `FileCache` class
can now store any value (not only HTTP responses). Therefor it's also
moved from the `Crwlr\Crawler\Loader\Http\Cache` namespace up to
`Crwlr\Crawler\Cache`. Further made the `HttpResponseCacheItem`
obsolete. Its functionality is split to the new general `CacheItem`
class and the `RespondedRequest` class. If you don't already provide an
instance of `CacheItem` to the `FileCache::set()` method, it wraps the
value you want to cache in a `CacheItem` before serializing it and
writing to a file, so it always contains the date when it was created
and the time to live, for the ttl implementation. This is all
implemented with backwards compatibility in mind. It will still work
with cache files generated with package versions before 1.0.