Replace dontCascade() with excludeFromGroupOutput()
After the change in v0.7 that groups can only produce combined output,
the dontCascade() functionality doesn't really make sense anymore to
be available on any step. So it's removed.
For the use case of excluding the output of one step from a groups
combined output, add a new method `excludeFromGroupOutput()` to the
`Step` class (so it's not available on a `Group` step instance).