Add Step::refineOutput() functionality
Add new method `Step::refineOutput()` to manually refine step output
values. It takes either a `Closure` or an instance of the new
`RefinerInterface` as argument. If the step produces array output, you
can provide a key from the array output, to refine, as first argument
and the refiner as second argument. You can call the method multiple
times and all the refiners will be applied to the outputs in the order
you add them. If you want to refine multiple output array keys with a
`Closure`, you can skip providing a key and the `Closure` will receive
the full output array for refinement. As mentioned you can provide an
instance of the `RefinerInterface`. There are already a few
implementations: `Str::afterFirst()`, `Str::afterLast()`,
`Str::beforeFirst()`, `Str::beforeLast()`, `Str::betweenFirst()`,
`Str::betweenLast()` and `Str::replace()`.