Links/URLs without fragment part
The `HttpCrawl` step (`Http::crawl()`) by default now removes the
fragment part of URLs to not load the same page multiple times, because
in almost any case, servers won't respond with different content based
on the fragment. That's why this change is considered non-breaking. For
the rare cases when servers respond with different content based on the
fragment, you can call the new `keepUrlFragment()` method of the step.

The `GetLink` (`Html::getLink()`) and `GetLinks` (`Html::getLinks()`)
steps, as well as the abstract `DomQuery` (parent of `CssSelector`
(/`Dom::cssSelector`) and `XPathQuery` (/`Dom::xPath`)) now have a
method `withoutFragment()` to get links respectively URLs without their
fragment part.