Crawl step can optionally use canonical links
The `HttpCrawl` step (`Http::crawl()`) has a new method
`useCanonicalLinks()`. If you call it, the step will not yield responses
if its canonical link URL was already yielded. And if it discovers a
link, and some document pointing to that URL via canonical link was
already loaded, it treats it as if it was already loaded. Further this
feature also sets the canonical link URL as the `effectiveUri` of the
response.