New method HttpLoader::cacheOnlyWhereUrl()
The method takes an instance of the `FilterInterface` as argument, so
users can control which responses will be cached (filtering by the
response URLs).

Besides that also upgrade PHP CS Fixer rules to PER and add
no_unused_imports rule. And change to running cs-fix instead of cs in
the git pre-commit hook.