Fix reading uncompressed files
When FileCache compression is activated, the FileCache assumed that all
cache files that it will read are compressed. But there could be
existing cache files created without compression activated which lead to
errors. The FileCache now first checks if the content is compressed and
only tries to decode if it is.