New DomQuery::formattedText() method
The `DomQuery` class (parent of `CssSelector` (`Dom::cssSelector`) and
`XPathQuery` (`Dom::xPath`)) has a new method `formattedText()` that
uses the new crwlr/html-2-text package to convert the HTML to formatted
plain text. You can also provide a customized instance of the
`Html2Text` class to the `formattedText()` method.

Also added a lot of missing @throws doblock tags in the whole codebase
and adapted PHP CS Fixer config a little.

Further, fixed an issue in the HttpCrawl step, if a newly found URL
redirects to a URL that was already loaded before, don't yield the
response, so we don't have the same page twice in the output.