Merge HttpBaseLoader back to HttpLoader
Merge `HttpBaseLoader` again with `HttpLoader`. It's probably not a good
idea to have multiple loaders. At least not multiple loaders just for
HTTP. It should be enough to publicly expose the
`HeadlessBrowserLoaderHelper` via `HttpLoader::browserHelper()` for the
extension steps. But keep the `HttpBase` step, to share the general HTTP
functionality implemented there.