Improve store call timing and memory usage
Improvement regarding the timing when a store (`Store` class instance)
is called by the crawler with a final crawling result. When a crawling
step initiates a crawling result (so, `addToResult()` was called on the
step instance), the crawler has to wait for all child outputs (resulting
from one step-input) until it calls the store, because the child outputs
can all add data to the same final result object. But previously this
was not only the case for all child outputs starting from a step where
`addToResult()` was called, but all children of one initial crawler
input. So with this change, in a lot of cases, the store will earlier be
called with finished `Result` objects and memory usage will be lowered.