Apply cache filter also for usage of cached responses
When using `HttpLoader::cacheOnlyWhereUrl()` to restrict caching, the
filter rule is not only applied when adding newly loaded responses to
the cache, but also for using cached responses. Example: a response for
`https://www.example.com/foo` is already available in the cache, but
`$loader->cacheOnlyWhereUrl(Filter::urlPathStartsWith('/bar/'))` was
called, the cached response is not used.