Fix getting non HTML documents via browser
In production getting the raw response body in the listener for
Network.responseReceived sometimes failed. Waiting a little helped.
So, as it seems that sometimes it is just not ready yet, move getting
the raw response body to a later point and also do it only when the
responseIsHtmlDocument() method returns false.