Screenshot action and Http step improvements
Added a screenshot browser action. Therefor extended the
`RespondedRequest` class, so it also transports the paths of the
screenshots on the filesystem.
Also added two new methods to the `Http` steps allowing to modify the
loader, only for one particular step in the chain. They can be used to
skip the configured cache or to switch to use the (headless) Chrome
browser.