Handle invalid characters in XML
As sometimes, XML parsing errors occur because of characters that aren't
valid within XML documents, the `XmlDocument` class now catches
exceptions, tries to find and replace invalid characters (with
transliterates or HTML entities) and retries parsing the document.
If installed, it tries to find transliterates for characters, using the
`voku/portable-ascii` composer package. Added the package to the
`suggest` section in the composer.json file.