Result::toArray() convert all objects to arrays
The `Result::toArray()` method now converts all objects contained in the
Result array (at any level of the array) to arrays. Also, if the only
element in a result array has some autogenerated key containing
"unnamed", but the value also is an associative array with string keys,
the method only returns that child array.